Remove the grpc stuff, tonic, tower, tonic-health etc.
