+++ iroh-gateway/Cargo.toml
@@ -68,5 +68,5 @@ tempfile = "3.3.0"
 
 [features]
 default = ["rpc-grpc", "rpc-mem"]
-rpc-grpc = ["iroh-rpc-types/grpc", "iroh-rpc-client/grpc", "iroh-metrics/rpc-grpc"]
-rpc-mem = ["iroh-rpc-types/mem", "iroh-rpc-client/mem"]
+rpc-grpc = ["iroh-metrics/rpc-grpc"]
+rpc-mem = []
