+++ iroh-rpc-client/Cargo.toml
@@ -12,7 +12,6 @@ rust-version = "1.63"
 
 anyhow = "1"
 async-stream = "0.3.3"
-async-trait = "0.1.56"
 bytes = "1.1.0"
 cid = "0.8.0"
 config = "0.13.1"
@@ -20,17 +19,12 @@ futures = "0.3.21"
 iroh-metrics = { path = "../iroh-metrics", default-features = false }
 iroh-util = { path = "../iroh-util" }
 libp2p = { version = "0.50", default-features = false, features = ["gossipsub"] }
-paste = "1.0.7"
-prost = "0.11"
 quic-rpc = { version = "0.1.4" }
 serde = { version = "1.0", features = ["derive"] }
 tokio = { version = "1", features = ["sync"] }
 toml = "0.5.9"
-tonic = { version = "0.8", optional = true }
-tonic-health = { version = "0.7", optional = true }
-tower = { version = "0.4.13", optional = true }
 tracing = "0.1.34"
-iroh-rpc-types = { path = "../iroh-rpc-types", optional = true }
+iroh-rpc-types = { path = "../iroh-rpc-types" }
 quinn = { version = "0.9.1" }
 rustls = { version = "0.20.7", features = ["dangerous_configuration"] }
 rcgen = "0.10.0"
@@ -41,5 +35,5 @@ tokio-stream = { version = "0.1.9", features = ["net"] }
 
 [features]
 default = []
-grpc = ["tonic", "tonic-health", "iroh-rpc-types/grpc", "iroh-rpc-types/mem", "iroh-metrics/rpc-grpc", "tower"]
-mem = ["iroh-rpc-types/mem"]
+grpc = ["iroh-metrics/rpc-grpc"]
+mem = []
