+++ iroh-rpc-types/Cargo.toml
@@ -10,40 +10,17 @@ rust-version = "1.63"
 
 [dependencies]
 anyhow = "1"
-async-trait = "0.1.56"
 bytes = { version = "1.3.0", features = ["serde"] }
 cid = { version = "0.8.6", features = ["serde-codec"] }
 derive_more = "0.99.17"
 futures = "0.3.24"
-iroh-metrics = { path = "../iroh-metrics", default-features = false }
 multiaddr = "0.16.0"
-paste = "1.0.7"
-prost = "0.11"
-prost-types = "0.11.1"
 quic-rpc = { version = "0.1.4" }
 serde = { version = "1" }
 serde_with = "2.0.0"
-tokio = { version = "1", features = ["net", "sync"] }
-tokio-stream = { version = "0.1.9", optional = true, features = ["net"] }
-tonic = { version = "0.8", optional = true }
-tonic-health = { version = "0.7", optional = true }
-
-[build-dependencies]
-prost-build = "0.11.1"
-tonic-build = { version = "0.8", features = ["prost"], optional = true }
 
 [features]
 default = []
-# GRPC based transport
-grpc = [
-  "tonic",
-  "tonic-build",
-  "tonic-health",
-  "iroh-metrics/rpc-grpc",
-  "tokio-stream"
-]
-# Memory based transport
-mem = []
 
 # Builds and empty tonic server & client as well as associated protobufs
 # that can be used for testing
