+++ iroh-rpc-types/src/gateway.rs
