--- iroh-rpc-types/src/grpc/gateway.rs
@@ -1,6 +1,0 @@
-include_proto!("gateway");
-
-proxy!(
-    Gateway,
-    version: () => VersionResponse => VersionResponse
-);
