--- iroh-rpc-types/src/grpc/mod.rs
@@ -1,13 +1,0 @@
-pub mod addr;
-pub mod gateway;
-pub mod p2p;
-pub mod store;
-
-#[cfg(feature = "testing")]
-pub mod test;
-
-pub use addr::Addr;
-
-// Reexport for convenience.
-#[cfg(feature = "grpc")]
-pub use tonic::transport::NamedService;
