--- iroh-rpc-types/src/grpc/test.rs
@@ -1,1 +1,0 @@
-include_proto!("test");
