+++ iroh-rpc-types/src/lib.rs
@@ -1,2 +1,10 @@
-pub mod qrpc;
-pub use qrpc::*;
+pub mod addr;
+pub mod gateway;
+pub mod p2p;
+pub mod store;
+
+pub use addr::Addr;
+
+pub trait NamedService {
+    const NAME: &'static str;
+}
