--- iroh-rpc-types/src/qrpc/mod.rs
@@ -1,10 +1,0 @@
-pub mod addr;
-pub mod gateway;
-pub mod p2p;
-pub mod store;
-
-pub use addr::Addr;
-
-pub trait NamedService {
-    const NAME: &'static str;
-}
