+++ iroh-rpc-types/src/store.rs
