+++ iroh-share/Cargo.toml
@@ -19,8 +19,8 @@ futures = "0.3.21"
 iroh-metrics = { path = "../iroh-metrics", default-features = false }
 iroh-p2p = { path = "../iroh-p2p", default-features = false, features = ["rpc-mem"] }
 iroh-resolver = { path = "../iroh-resolver", default-features = false, features = ["rpc-mem"] }
-iroh-rpc-client = { path = "../iroh-rpc-client", default-features = false, features = ["mem"] }
-iroh-rpc-types = { path = "../iroh-rpc-types", default-features = false, features = ["mem"] }
+iroh-rpc-client = { path = "../iroh-rpc-client", default-features = false }
+iroh-rpc-types = { path = "../iroh-rpc-types", default-features = false }
 iroh-store = {path = "../iroh-store", default-features = false, features = ["rpc-mem"] }
 iroh-util = { path = "../iroh-util" }
 libp2p = { version = "0.50", default-features = false, features = ["gossipsub"] }
