+++ iroh-store/Cargo.toml
@@ -45,8 +45,8 @@ tokio = { version = "1", features = ["rt", "macros", "rt-multi-thread"] }
 
 [features]
 default = ["rpc-grpc", "rpc-mem"]
-rpc-grpc = ["iroh-rpc-types/grpc", "iroh-rpc-client/grpc", "iroh-metrics/rpc-grpc"]
-rpc-mem = ["iroh-rpc-types/mem", "iroh-rpc-client/mem"]
+rpc-grpc = ["iroh-metrics/rpc-grpc"]
+rpc-mem = []
 
 [[bench]]
 name = "store"
