+++ iroh/Cargo.toml
@@ -29,7 +29,6 @@ relative-path = { version = "1.7.2", optional = true }
 serde = { version = "1.0", features = ["derive"] }
 sysinfo = "0.26.4"
 tokio = { version = "1", features = ["fs", "io-util"] }
-tonic = "0.8"
 tracing = "0.1.34"
 which = "4.3.0"
